+++ src/Form/Extension/ExtendedChannelExtension.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace MangoSylius\ExtendedChannelsPlugin\Form\Extension;
 
 use MangoSylius\ExtendedChannelsPlugin\Entity\TimezoneEntity;
-use Sylius\Bundle\ChannelBundle\Form\Type\ChannelType;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
@@ -39,11 +38,11 @@ final class ExtendedChannelExtension extends AbstractTypeExtension
 			]);
 	}
 
-	/**
-	 * {@inheritdoc}
-	 */
-	public function getExtendedType(): string
+	/** @return array<int, string> */
+	public static function getExtendedTypes(): array
 	{
-		return ChannelType::class;
+		return [
+			\Sylius\Bundle\ChannelBundle\Form\Type\ChannelType::class,
+		];
 	}
 }