+++ src/Form/Type/HelloBarTranslationType.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MangoSylius\ExtendedChannelsPlugin\Form\Type;
+
+use Sylius\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+
+final class HelloBarTranslationType extends AbstractResourceType
+{
+	/**
+	 * {@inheritdoc}
+	 */
+	public function buildForm(FormBuilderInterface $builder, array $options): void
+	{
+		$builder
+			->add('title', TextType::class, ['required' => false])
+			->add('content', TextareaType::class, ['required' => false]);
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function getBlockPrefix(): string
+	{
+		return 'hello_bar_translation';
+	}
+}